Integrate bpm:setup-email command into bpm:install
